--- app/src/main/java/com/example/hdmedi/activity/BindingActivity.kt
@@ -1,18 +1,0 @@
-package com.example.hdmedi.activity
-
-import android.os.Bundle
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-
-abstract class BindingActivity<T : ViewDataBinding>(
-    @LayoutRes private val layoutResId: Int
-) : AppCompatActivity() {
-    protected lateinit var binding: T
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, layoutResId)
-    }
-}
